Remove side effects from import of denoflate
This stops the import of "https://deno.land/x/denoflate/mod.ts" having
the side effect of downloading
https://deno.land/x/denoflate@1.1/pkg/denoflate_bg.wasm because ideally
imports do not have any side effects, see
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/import

This is a breaking change for users of the module who will have to
explicitly call `initWasm()` before making use of any of the available
functions.

Making this change enables users of the module to decide if/when to run
this initialisation.